# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SForum API
  version: 0.1.0
  description: Initial API contract for the SForum foundation.
servers:
  - url: /api/v1
paths:
...
  /permissions:
    get:
      summary: List global permissions
      operationId: listPermissions
      responses:
        "200":
          description: Permission catalog visible to role or user managers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePermissionList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /permissions/matrix:
    get:
      summary: List role permission matrix
      operationId: listPermissionMatrix
      responses:
        "200":
          description: Permission catalog with permissions assigned to each role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponsePermissionMatrix"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /roles:
    get:
      summary: List roles and user groups
      operationId: listRoles
      responses:
        "200":
          description: Roles visible to role managers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRoleList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      summary: Create a custom role
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "201":
          description: Role created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRole"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /roles/{roleKey}:
    patch:
      summary: Update a role alias and description
      operationId: updateRole
      parameters:
        - $ref: "#/components/parameters/RoleKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "200":
          description: Role updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseRole"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
    delete:
      summary: Delete a custom role
      operationId: deleteRole
      parameters:
        - $ref: "#/components/parameters/RoleKey"
      responses:
        "200":
          description: Role deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseNull"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /roles/{roleKey}/permissions:
    put:
      summary: Replace permissions assigned to a role
      operationId: replaceRolePermissions
      parameters:
        - $ref: "#/components/parameters/RoleKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - permissions
              properties:
                permissions:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Role permissions replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseNull"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /users/{userID}:
    get:
      summary: Get an admin user detail
      operationId: getAdminUser
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: User detail with effective permissions and direct overrides.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAdminUserDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /users/{userID}/roles:
    put:
      summary: Replace roles assigned to a user
      operationId: replaceUserRoles
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRolesInput"
      responses:
        "200":
          description: User roles replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAdminUserDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /users/{userID}/permission-overrides:
    put:
      summary: Replace direct permission overrides for a user
      operationId: replaceUserPermissionOverrides
      parameters:
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPermissionOverridesInput"
      responses:
        "200":
          description: User direct permission overrides replaced.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAdminUserDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
components:
  parameters:
    RoleKey:
      name: roleKey
      in: path
      required: true
      schema:
        type: string
      description: Stable role key, such as member or moderator.
    UserID:
      name: userID
      in: path
      required: true
      schema:
        type: integer
        format: int64
      description: Numeric user id.
  responses:
    Unauthorized:
      description: Authentication is required or credentials are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Forbidden:
      description: The current user does not have permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Conflict:
      description: The request conflicts with a system invariant.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    UnprocessableEntity:
      description: The request body is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
  schemas:
    ApiErrorData:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          description: Stable machine-readable reason code, such as auth.register_invalid.
        fields:
          type: object
          description: Localized field-level messages keyed by request field name. Registration uses username, email, password, and humanVerification.
          additionalProperties:
            type: array
            items:
              type: string
    ApiErrorResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/ApiErrorData"
    ApiResponseNull:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          examples: [200]
        message:
          type: string
          examples: [OK]
        data:
          type: "null"
...
    ApiResponseRole:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/Role"
    ApiResponseRoleList:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/Role"
    ApiResponsePermissionList:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
    ApiResponsePermissionMatrix:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/PermissionMatrix"
...
    ApiResponseAdminUserDetail:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: "#/components/schemas/AdminUserDetail"
...
    Role:
      type: object
      required:
        - id
        - key
        - alias
        - description
        - isSystem
        - isDefault
        - isDeletable
        - isEnabled
        - permissionKeys
      properties:
        id:
          type: integer
          format: int64
        key:
          type: string
        alias:
          type: string
        description:
          type: string
        isSystem:
          type: boolean
        isDefault:
          type: boolean
        isDeletable:
          type: boolean
        isEnabled:
          type: boolean
        permissionKeys:
          type: array
          items:
            type: string
    RoleInput:
      type: object
      properties:
        key:
          type: string
        alias:
          type: string
        description:
          type: string
    Permission:
      type: object
      required:
        - key
        - module
        - description
      properties:
        key:
          type: string
        module:
          type: string
        description:
          type: string
    RolePermissionSet:
      type: object
      required:
        - roleKey
        - permissionKeys
      properties:
        roleKey:
          type: string
        permissionKeys:
          type: array
          items:
            type: string
    PermissionMatrix:
      type: object
      required:
        - permissions
        - roles
      properties:
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RolePermissionSet"
    PermissionOverrides:
      type: object
      required:
        - allow
        - deny
      properties:
        allow:
          type: array
          items:
            type: string
        deny:
          type: array
          items:
            type: string
    AdminUserSummary:
      type: object
      required:
        - id
        - username
        - email
        - displayName
        - locale
        - status
        - isInitialSuperAdmin
        - roleKeys
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        locale:
          type: string
        status:
          type: string
          enum: [active, disabled, banned]
        isInitialSuperAdmin:
          type: boolean
        roleKeys:
          type: array
          items:
            type: string
    AdminUserDetail:
      allOf:
        - $ref: "#/components/schemas/AdminUserSummary"
        - type: object
          required:
            - permissions
            - permissionOverrides
          properties:
            permissions:
              type: array
              items:
                type: string
            permissionOverrides:
              $ref: "#/components/schemas/PermissionOverrides"
...
    UserRolesInput:
      type: object
      required:
        - roleKeys
      properties:
        roleKeys:
          type: array
          items:
            type: string
    UserPermissionOverridesInput:
      $ref: "#/components/schemas/PermissionOverrides"